requirements.txt
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
bidict==0.21.3
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==3.4.7
decorator==5.1.1
distlib==0.3.3
Django==3.2.6
easyocr==1.4.1
executing==0.8.3
filelock==3.3.0
Flask==0.10.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
ipython==8.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.26
ndg-httpsclient==0.5.1
networkx==2.7.1
numpy==1.21.3
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.3.4
paramiko==2.7.2
parso==0.8.3
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.4.0
prompt-toolkit==3.0.29
pure-eval==0.2.2
pusher==1.2.3
pyasn1==0.4.8
pycparser==2.20
Pygments==2.11.2
pygorithm==1.0.4
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
Pysher @ git+https://github.com/nlsdfnbch/Pysher.git@8515f9db9a4dc995b012648c1aae4bf13334ba61
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.26.0
scikit-image==0.19.2
scipy==1.8.0
shadowsocks==2.8.2
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
sshtunnel==0.4.0
stack-data==0.2.0
termcolor==1.1.0
tifffile==2022.3.25
torch==1.11.0
torchvision==0.12.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.7
virtualenv==20.8.1
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.2
ws2s-python==2.1.5
zope.event==4.5.0
zope.interface==5.4.0
